refactor: electron from preload into seperate file
